extracted logic to wrapper & renamed 2D surface plot file
